Move getDelegates into accounts.js
