+++ services/core/shared/core/compat/sdk_v5/index.js
@@ -24,6 +24,7 @@ const {
 
 const {
     getAccounts,
+    getDelegates,
     getMultisignatureGroups,
     getMultisignatureMemberships,
     validateAddress,
@@ -60,10 +61,6 @@ const {
     getCachedAccountBySecondPublicKey,
     getCachedAccountByUsername,
 } = require('./coreCache');
-
-const {
-    getDelegates,
-} = require('./delegates');
 
 const {
     getEstimateFeeByteForBatch,
